+++ src/HIndent/Pretty/Combinators/Lineup.hs
@@ -243,7 +243,7 @@ vWrappedLineup sep (prefix, suffix) ps =
     >> space |=> do
          prefixedLined [sep, ' '] ps
          newline
-         indentedWithSpace (-(fromIntegral (length prefix) + 1)) $ string suffix
+         indentedWithSpace (-(length prefix + 1)) $ string suffix
 
 -- | Similar to 'vWrappedLineup' but the suffix is in the same line as the
 -- last element.
